backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
PyMuPDF
pydantic